BOAC-1468: Vueifies curated group sort and search
