+++ boac/static/app/main.css
@@ -219,6 +219,7 @@
   font-size: 24px;
   font-weight: bold;
   margin-bottom: 10px;
+  margin-top: 20px;
 }
 
 .page-section-header-sub {
