+++ src/components/curated/CuratedGroupStudent.vue
@@ -1,0 +1,15 @@
+<template>
+  <div class="cohort-column-results">
+    List
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'CuratedGroupList',
+  props: ['curatedGroup']
+};
+</script>
+
+<style scoped>
+</style>
